Move component base classes into a single file (#8918)
